requirements.txt
Django == 4.1.1
djangorestframework
PyJWT
Pillow == 9.2.0
requests == 2.28.1